requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.0
openai==1.32.0
pyyaml==5.4.1
scapy==2.5.0
pyshark==0.4.4
faker==18.9.0
jinja2==3.1.2
matplotlib==3.8.0
nltk==3.8.1
pandas==2.2.2
scikit-learn==1.3.2
pytest==7.4.0